+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## 1.0.0
 
+- `extension` for `int`, `double` and `num` collections.
 - Initial version.
-- `extension` for `int`, `double` and `num`.